Respond to final PR comments. Move ExperimentalPlugin to its own file.
